Minor refactor to avoid large headers
Just expose single public functions, e.g. "RunTypeApplicator", so all the
visitors don't have to be defined in the header files. This is specifically
for visitors used in type_resolver.cpp.

No functional changes.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/5021, branch: user/jordalgo/type_reorganize/3
