+++ src/ast/passes/types/cast_creator.h
@@ -1,16 +1,18 @@
 #pragma once
 
-#include "ast/pass_manager.h"
-#include "ast/visitor.h"
-
 #include <optional>
-#include <variant>
+
+#include "types.h"
 
 namespace bpftrace {
 class BPFtrace;
 } // namespace bpftrace
 
 namespace bpftrace::ast {
+
+class ASTContext;
+class Expression;
+struct Program;
 
 std::optional<SizedType> try_tuple_cast(ASTContext &ctx,
                                         Expression &exp,
@@ -22,28 +24,6 @@ std::optional<SizedType> try_record_cast(ASTContext &ctx,
                                          const SizedType &expr_type,
                                          const SizedType &target_type);
 
-class CastCreator : public Visitor<CastCreator> {
-public:
-  explicit CastCreator(ASTContext &ast, BPFtrace &bpftrace);
-
-  using Visitor<CastCreator>::visit;
-  void visit(AssignMapStatement &assignment);
-  void visit(AssignVarStatement &assignment);
-  void visit(Binop &binop);
-  void visit(BlockExpr &block);
-  void visit(Call &call);
-  void visit(Cast &cast);
-  void visit(For &f);
-  void visit(IfExpr &if_expr);
-  void visit(Jump &jump);
-  void visit(MapAccess &acc);
-  void visit(Probe &probe);
-  void visit(Subprog &subprog);
-
-private:
-  ASTContext &ctx_;
-  BPFtrace &bpftrace_;
-  std::variant<std::monostate, Probe *, Subprog *> top_level_node_;
-};
+void RunCastCreator(ASTContext &ast, BPFtrace &bpftrace);
 
 } // namespace bpftrace::ast
